Add LICENSE and README
Those are the Quake II maptools, fixed to work on unixoid platforms.
This is work in progress and unfinished, the code may or may not work.
Use this code at your own risk and do not complain if it breaks in
interesting ways. Patches (or even better pull requests) and success
reports are highly welcome.